+++ cogs/gatekeeper/gatekeeper.py
@@ -71,71 +71,6 @@ class Gatekeeper(rubbercog.Rubbercog):
             delete_after=config.get("delay", "verify"),
         )
 
-    @commands.check(check.is_in_quarantine_or_dm)
-    @commands.check(check.is_quarantined)
-    @commands.cooldown(rate=5, per=120, type=commands.BucketType.user)
-    @commands.group(name="reverify")
-    async def reverify(self, ctx):
-        await utils.send_help(ctx)
-
-    @reverify.command(name="change", aliases=["update", "downgrade", "upgrade"])
-    async def reverify_change(self, ctx, email: str):
-        """Change your e-mail
-
-        email: Your new verification e-mail
-        """
-        await utils.delete(ctx)
-
-        if "@" not in email or len(email.split("@")) > 2:
-            raise NotAnEmail()
-
-        # check if user is in database
-        if repo_u.get(ctx.author.id) is None:
-            raise NotInDatabase()
-
-        # check the database for email
-        if repo_u.getByLogin(email) is not None:
-            raise EmailAlreadyInDatabase()
-
-        role = await self._email_to_role(ctx, email)
-        repo_u.update(discord_id=ctx.author.id, group=role.name)
-
-        # TODO To text hjson
-        await self.output.info(ctx, "Mail changed")
-
-    @reverify.command(name="prove")
-    async def reverify_prove(self, ctx):
-        """Ask for verification code"""
-        await utils.delete(ctx)
-
-        db_user = repo_u.get(ctx.author.id)
-
-        if db_user is None:
-            await self.console.error("User in `?reverify prove` is not in database.")
-            raise NotInDatabase()
-
-        if db_user.status != "quarantined":
-            await self.console.error("User in `?reverify prove` did not have `quarantined` status.")
-            raise UnexpectedReverify()
-
-        if db_user.group == "FEKT" and "@" not in db_user.login:
-            email = db_user.login + "@stud.feec.vutbr.cz"
-        elif db_user.group == "VUT" and "@" not in db_user.login:
-            email = db_user.login + "@vutbr.cz"
-        else:
-            email = db_user.login
-
-        # generate new code
-        code = await self._update_user(ctx.author)
-
-        # send mail
-        await self._send_verification_email(ctx.author, email, code)
-
-        await ctx.send(
-            self.text.get("reverify successful", mention=ctx.author.mention),
-            delete_after=config.get("delay", "verify"),
-        )
-
     @commands.check(check.is_not_verified)
     @commands.cooldown(rate=3, per=120, type=commands.BucketType.user)
     @commands.command()
@@ -145,11 +80,7 @@ class Gatekeeper(rubbercog.Rubbercog):
 
         db_user = repo_u.get(ctx.author.id)
 
-        if (
-            db_user is None
-            or db_user.status in ("unknown", "unverified", "quarantined")
-            or db_user.code is None
-        ):
+        if db_user is None or db_user.status in ("unknown", "unverified") or db_user.code is None:
             raise SubmitWithoutCode()
 
         if db_user.status != "pending":
@@ -164,12 +95,7 @@ class Gatekeeper(rubbercog.Rubbercog):
         repo_u.save_verified(ctx.author.id)
 
         # add role
-        unverified = await self._add_verify_roles(ctx.author, db_user)
-        if unverified:
-            return await ctx.send(
-                self.text.fill("reverification public", mention=ctx.author.mention),
-                delete_after=config.get("delay", "verify"),
-            )
+        await self._add_verify_roles(ctx.author, db_user)
 
         # send messages
         for role_id in config.get("roles", "native"):
@@ -188,12 +114,29 @@ class Gatekeeper(rubbercog.Rubbercog):
         # fmt: on
 
         await self.event.user(
-            ctx, f"User {ctx.author.id} verified with group **{db_user.group}**.",
+            ctx,
+            f"User {ctx.author.id} verified with group **{db_user.group}**.",
         )
+
+    ##
+    ## Listeners
+    ##
+
+    @commands.Cog.listener()
+    async def on_member_join(self, member: discord.Member):
+        """Add them their roles back, if they have been verified before"""
+        db_user = repo_u.get(member.id)
+        if db_user is None or db_user.status != "verified":
+            return
+
+        # user has been verified, give them their main roles back
+        await self._add_verify_roles(member, db_user)
+        await self.event.user(member, f"Verification skipped (**{db_user.group}**)")
 
     ##
     ## Helper functions
     ##
+
     async def _email_to_role(self, ctx, email: str) -> discord.Role:
         """Get role from email address"""
         registered = self.config.get("suffixes")
@@ -293,19 +236,12 @@ class Gatekeeper(rubbercog.Rubbercog):
             server.login(self.config.get("email", "address"), self.config.get("email", "password"))
             server.send_message(msg)
 
-    async def _add_verify_roles(self, member: discord.Member, db_user: object) -> bool:
+    async def _add_verify_roles(self, member: discord.Member, db_user: object):
         """Return True if reverified"""
         verify = self.getVerifyRole()
-        quarantine = self.getGuild().get_role(config.get("roles", "quarantine_id"))
         group = discord.utils.get(self.getGuild().roles, name=db_user.group)
 
-        unverified = False
-        if quarantine in member.roles:
-            await member.remove_roles(quarantine, reason="Reverification")
-            unverified = True
-
         await member.add_roles(verify, group, reason="Verification")
-        return unverified
 
     ##
     ## Error catching