+++ rubbergoddess.py
@@ -1,6 +1,7 @@
 import traceback
 from datetime import datetime
 
+import discord
 from discord.ext import commands
 
 from core import help, rubbercog, output, presence, utils
@@ -16,7 +17,11 @@ from repository.database.points import Points
 from repository.database.acl import ACL_group, ACL_rule, ACL_rule_user, ACL_rule_group
 from repository.review_repo import ReviewRepository
 
-bot = commands.Bot(command_prefix=config.prefix, help_command=help.Help())
+bot = commands.Bot(
+    command_prefix=config.prefix,
+    help_command=help.Help(),
+    allowed_mentions=discord.AllowedMentions(roles=False, everyone=False, users=True),
+)
 
 presence = presence.Presence(bot)
 event = output.Event(bot)
@@ -33,17 +38,18 @@ started = False
 
 @bot.event
 async def on_ready():
-    """If Rubbergoddess is ready"""
     global started
     if started:
-        return
-    started = True
-
-    message = "Logged in [{level}]: {timestamp} (hash {hash})".format(
-        level=config.get("bot", "logging"),
-        timestamp=datetime.today().strftime("%Y-%m-%d %H:%M:%S"),
-        hash=utils.git_hash()[:7],
-    )
+        message = "Reconnected: {timestamp}".format(
+            timestamp=datetime.today().strftime("%Y-%m-%d %H:%M:%S")
+        )
+    else:
+        message = "Logged in [{level}]: {timestamp} (hash {hash})".format(
+            level=config.get("bot", "logging"),
+            timestamp=datetime.today().strftime("%Y-%m-%d %H:%M:%S"),
+            hash=utils.git_hash()[:7],
+        )
+        started = True
 
     print(message)
     channel = bot.get_channel(config.get("channels", "stdout"))